Fix: Unintentional JVM Options
Enclose the APP_HOME variable expansion in double quotes to
properly handle values containing spaces. Without quotes,
"default_jvm_opts" is interpreted as multiple options split at the
spaces, leading to script failure or unintended JVM options being
used.

"sdklib/BUILD" and "misc/screenshot2/BUILD" also contain lint
violations / errors. "java_library" needs to be loaded in both
files. In "sdklib/BUILD", "combine_licenses" needs to be removed
because it is not used and the load statement for "jvm_import"
needs to be at the top of the file with other load statements.

Change-Id: Ic4d70ca2b9a7fb64779d11bcae2726141a2fd5b0