Don't warn about implicit new instances of value/data classes
The Lint issue ImplicitSamInstance warns when passing newly constructed instances to functions like `removeListener()`. That gives a false positive when the listener has equality implemented, where removals based on equality will work. Example: https://pl.kotl.in/-vPWLhMfkZ

This relaxes the detector to not warn if the type being constructed is a `value class` or `data class`, as those are guaranteed to have working equality.

Change-Id: I42b8551ff59de875111dfa41698df710c53e42d5