Disable interpreting unknown minSdk versions as the latest SDK.
This generates a lot of false positives in projects that use placeholder variables to configure minSdk (yaqs/2517571565328007168, b/262911421, b/306194463). It's bad to assume that "unknown" means "latest", especially the case of minSdk, because almost no project would use the latest SDK as its minSdk. We shouldn't be generating warnings based on this bad assumption, suggesting that version checks are unnecessary when they are in fact necessary.

Change-Id: I01c9bbd73ec809a7248798a20b94750e8a248803